# packages/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stacks Mesh API
  description: A Mesh API (formerly Rosetta API) implementation for the Stacks blockchain
  version: 0.7.9
...
paths:
...
  /construction/derive:
    post:
      tags:
        - Construction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - network_identifier
                - public_key
              properties:
                network_identifier:
                  type: object
                  required:
                    - blockchain
                    - network
                  properties:
                    blockchain:
                      type: string
                      enum:
                        - stacks
                    network:
                      anyOf:
                        - type: string
                          enum:
                            - mainnet
                        - type: string
                          enum:
                            - testnet
                public_key:
                  type: object
                  required:
                    - hex_bytes
                    - curve_type
                  properties:
                    hex_bytes:
                      pattern: ^(0x)?[0-9a-fA-F]+$
                      type: string
                    curve_type:
                      type: string
                      enum:
                        - secp256k1
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                  account_identifier:
                    type: object
                    required:
                      - address
                    properties:
                      address:
                        anyOf:
                          - pattern: ^[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{28,41}\.[a-zA-Z]([a-zA-Z0-9]|[-_]){0,39}$
                            type: string
                          - pattern: ^[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{28,41}$
                            type: string
                      sub_account:
                        type: object
                        required:
                          - address
                        properties:
                          address:
                            type: string
                          metadata:
                            type: object
                            additionalProperties: {}
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - retriable
                properties:
                  code:
                    type: number
                  message:
                    type: string
                  description:
                    type: string
                  retriable:
                    type: boolean
                  details:
                    type: object
                    additionalProperties: {}
...
  /construction/combine:
    post:
      tags:
        - Construction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - network_identifier
                - unsigned_transaction
                - signatures
              properties:
                network_identifier:
                  type: object
                  required:
                    - blockchain
                    - network
                  properties:
                    blockchain:
                      type: string
                      enum:
                        - stacks
                    network:
                      anyOf:
                        - type: string
                          enum:
                            - mainnet
                        - type: string
                          enum:
                            - testnet
                unsigned_transaction:
                  pattern: ^(0x)?[0-9a-fA-F]+$
                  type: string
                signatures:
                  minItems: 1
                  type: array
                  items:
                    type: object
                    required:
                      - signing_payload
                      - public_key
                      - signature_type
                      - hex_bytes
                    properties:
                      signing_payload:
                        type: object
                        required:
                          - account_identifier
                          - hex_bytes
                          - signature_type
                        properties:
                          address:
                            anyOf:
                              - pattern: ^[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{28,41}\.[a-zA-Z]([a-zA-Z0-9]|[-_]){0,39}$
                                type: string
                              - pattern: ^[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{28,41}$
                                type: string
                          account_identifier:
                            type: object
                            required:
                              - address
                            properties:
                              address:
                                anyOf:
                                  - pattern: ^[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{28,41}\.[a-zA-Z]([a-zA-Z0-9]|[-_]){0,39}$
                                    type: string
                                  - pattern: ^[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{28,41}$
                                    type: string
                              sub_account:
                                type: object
                                required:
                                  - address
                                properties:
                                  address:
                                    type: string
                                  metadata:
                                    type: object
                                    additionalProperties: {}
                          hex_bytes:
                            pattern: ^(0x)?[0-9a-fA-F]+$
                            type: string
                          signature_type:
                            type: string
                            enum:
                              - ecdsa_recovery
                      public_key:
                        type: object
                        required:
                          - hex_bytes
                          - curve_type
                        properties:
                          hex_bytes:
                            pattern: ^(0x)?[0-9a-fA-F]+$
                            type: string
                          curve_type:
                            type: string
                            enum:
                              - secp256k1
                      signature_type:
                        type: string
                        enum:
                          - ecdsa_recovery
                      hex_bytes:
                        pattern: ^(0x)?[0-9a-fA-F]+$
                        type: string
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - signed_transaction
                properties:
                  signed_transaction:
                    pattern: ^(0x)?[0-9a-fA-F]+$
                    type: string
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                  - retriable
                properties:
                  code:
                    type: number
                  message:
                    type: string
                  description:
                    type: string
                  retriable:
                    type: boolean
                  details:
                    type: object
                    additionalProperties: {}
...
servers:
  - url: http://localhost:3000
    description: Local
...