# packages/api/src/api/routes/construction.ts
import type { FastifyPluginAsyncTypebox } from '@fastify/type-provider-typebox';
import type { ApiConfig } from '../index.js';
import {
  getAddressFromPublicKey,
  makeUnsignedSTXTokenTransfer,
  deserializeTransaction,
  serializeTransaction,
  createMessageSignature,
  isSingleSig,
  sigHashPreSign,
  AuthType,
  makeUnsignedContractCall,
  hexToCV,
  makeUnsignedContractDeploy,
  StacksTransactionWire,
} from '@stacks/transactions';
import codec from '@stacks/codec';
import {
  ConstructionDeriveRequestSchema,
  ConstructionDeriveResponseSchema,
  ConstructionPreprocessRequestSchema,
  ConstructionPreprocessResponseSchema,
  ConstructionMetadataRequestSchema,
  ConstructionMetadataResponseSchema,
  ConstructionPayloadsRequestSchema,
  ConstructionPayloadsResponseSchema,
  ConstructionCombineRequestSchema,
  ConstructionCombineResponseSchema,
  ConstructionParseRequestSchema,
  ConstructionParseResponseSchema,
  ConstructionHashRequestSchema,
  ConstructionSubmitRequestSchema,
  TransactionIdentifierResponseSchema,
  ErrorResponseSchema,
} from '@stacks/mesh-schemas';
import type {
  ConstructionDeriveResponse,
  ConstructionParseResponse,
  ConstructionOptions,
  ConstructionOperation,
} from '@stacks/mesh-schemas';
import { STX_CURRENCY } from '../../utils/constants.js';
import { estimateTransactionFee, reorderSignatureToVrs } from '../../utils/construction.js';
import { MeshErrors } from '../../utils/errors.js';
import {
  addHexPrefix,
  removeHexPrefix,
  serializeDecodedTransactionOperations,
} from '../../serializers/index.js';
import BigNumber from 'bignumber.js';
import { isDeepStrictEqual } from 'node:util';

export const ConstructionRoutes: FastifyPluginAsyncTypebox<ApiConfig> = async (fastify, config) => {
  // `network` carries the node's actual chain ID (folded into the signing sighash), so
  // transactions are constructed/signed correctly even against a node running a custom chain ID.
  // Available in every mode; `rpcClient` (used by the node-backed endpoints below) is not.
  const { network } = config;

  fastify.post(
    '/construction/derive',
    {
      schema: {
        body: ConstructionDeriveRequestSchema,
        tags: ['Construction'],
        response: {
          200: ConstructionDeriveResponseSchema,
          500: ErrorResponseSchema,
        },
      },
    },
    async (request, reply) => {
      const { public_key } = request.body;
      try {
        const address = getAddressFromPublicKey(removeHexPrefix(public_key.hex_bytes), network);
        const response: ConstructionDeriveResponse = {
          account_identifier: { address },
        };
        return reply.send(response);
      } catch (error) {
        const message = error instanceof Error ? error.message : 'Unknown error';
        return reply.status(500).send(MeshErrors.invalidPublicKey(message));
      }
    }
  );

  fastify.post(
    '/construction/preprocess',
    {
      schema: {
        body: ConstructionPreprocessRequestSchema,
        tags: ['Construction'],
        response: {
          200: ConstructionPreprocessResponseSchema,
          500: ErrorResponseSchema,
        },
      },
    },
    async (request, reply) => {
      const { operations, metadata, max_fee, suggested_fee_multiplier } = request.body;
      try {
        const options = buildConstructionOptionsFromOperations({
          operations,
          memo: metadata?.memo ?? undefined,
          maxFee: max_fee?.[0]?.value,
          suggestedFeeMultiplier: suggested_fee_multiplier,
        });
        return reply.send({ options, required_public_keys: [{ address: options.sender_address }] });
      } catch (error) {
        return reply.status(500).send(MeshErrors.invalidTransaction((error as Error).message));
      }
    }
  );

  // Node-backed (fetches nonce/balance) — only served in online mode.
  if (config.mode === 'online') {
    const { rpcClient } = config;
    fastify.post(
      '/construction/metadata',
      {
        schema: {
          body: ConstructionMetadataRequestSchema,
          tags: ['Construction'],
          response: {
            200: ConstructionMetadataResponseSchema,
            500: ErrorResponseSchema,
          },
        },
      },
      async (request, reply) => {
        const { options, public_keys } = request.body;

        // Estimate the fee from the full serialized transaction size at a base per-byte rate, with
        // a floor (see `estimateTransactionFee`).
        //
        // TODO: this is a size-based estimate. The node's fee endpoints (`/v2/fees/transaction`,
        // `/v2/fees/transfer`) could give network-aware rates, but historically returned inaccurate
        // values, so we keep a self-contained estimate for now.
        // Fee = full serialized tx byte length × the base rate. If a dummy tx can't be built/sized
        // (unexpected op type or a build error), fall back to the configured default fee.
        let suggestedFee = config.constructionDefaultFee;
        const dummyPubKey = removeHexPrefix(public_keys?.[0]?.hex_bytes ?? '0'.repeat(66));
        try {
          switch (options.type) {
            case 'token_transfer': {
              const dummyTx = await makeUnsignedSTXTokenTransfer({
                recipient: options.recipient_address,
                amount: 1,
                fee: 0,
                nonce: 0,
                publicKey: dummyPubKey,
                network,
              });
              suggestedFee = estimateTransactionFee(dummyTx);
              break;
            }
            case 'contract_call': {
              // TODO: Should we check if the contract exists?
              const [contractAddress, contractName] = options.contract_identifier.split('.');
              const dummyTx = await makeUnsignedContractCall({
                contractAddress,
                contractName,
                functionName: options.function_name,
                functionArgs: options.args.map(arg => hexToCV(addHexPrefix(arg))),
                publicKey: dummyPubKey,
                fee: 0,
              });
              suggestedFee = estimateTransactionFee(dummyTx);
              break;
            }
            case 'contract_deploy': {
              const dummyTx = await makeUnsignedContractDeploy({
                contractName: options.contract_name,
                codeBody: options.source_code,
                clarityVersion: options.clarity_version,
                publicKey: dummyPubKey,
                fee: 0,
              });
              suggestedFee = estimateTransactionFee(dummyTx);
              break;
            }
          }
        } catch {
          // Couldn't build/serialize a dummy transaction to size the fee — use the default.
          suggestedFee = config.constructionDefaultFee;
        }
        // Apply fee multiplier if specified
        const feeMultiplier = options?.suggested_fee_multiplier
          ? Number(options.suggested_fee_multiplier)
          : undefined;
        if (feeMultiplier !== undefined) {
          suggestedFee = Math.round(suggestedFee * feeMultiplier);
        }
        // Cap fee if max_fee was specified in options
        const maxFee = options?.max_fee ? Number(options.max_fee) : undefined;
        if (maxFee !== undefined && suggestedFee > maxFee) {
          suggestedFee = maxFee;
        }

        const senderInfo = await rpcClient.request('GET', '/v2/accounts/{principal}', {
          params: { path: { principal: options.sender_address } },
        });
        return reply.send({
          metadata: {
            options,
            sender_account_info: {
              nonce: senderInfo.nonce,
              balance: BigNumber(senderInfo.balance).toString(),
            },
          },
          suggested_fee: [
            {
              value: String(suggestedFee),
              currency: STX_CURRENCY,
            },
          ],
        });
      }
    );
  }

  fastify.post(
    '/construction/payloads',
    {
      schema: {
        body: ConstructionPayloadsRequestSchema,
        tags: ['Construction'],
        response: {
          200: ConstructionPayloadsResponseSchema,
          500: ErrorResponseSchema,
        },
      },
    },
    async (request, reply) => {
      const { operations, metadata, public_keys } = request.body;
      const normalizedPublicKey = removeHexPrefix(public_keys[0].hex_bytes);

      // Build and validate construction options from operations.
      let options: ConstructionOptions;
      try {
        options = buildConstructionOptionsFromOperations({ operations });
      } catch (error) {
        return reply.status(500).send(MeshErrors.invalidTransaction((error as Error).message));
      }

      // The memo may arrive on the operations, at the top level of the request metadata (legacy
      // Rosetta style), or inside the metadata options (when it was given to
      // /construction/preprocess). Operation-level memos take precedence; conflicting sources are
      // rejected. The memo is resolved before the options equality check below so a memo that
      // rides in only via metadata does not fail the comparison.
      const operationMemo = options.type === 'token_transfer' ? options.memo : undefined;
      const metadataMemo =
        metadata.memo ??
        (metadata.options.type === 'token_transfer' ? metadata.options.memo : undefined);
      if (
        operationMemo !== undefined &&
        metadataMemo !== undefined &&
        operationMemo !== metadataMemo
      ) {
        return reply
          .status(500)
          .send(
            MeshErrors.invalidTransaction(
              'Memo from operations does not match the memo in metadata'
            )
          );
      }
      if (options.type === 'token_transfer') {
        options.memo = operationMemo ?? metadataMemo ?? undefined;
      }

      if (!isUnorderedDeepStrictEqual(withoutMemo(options), withoutMemo(metadata.options))) {
        return reply
          .status(500)
          .send(
            MeshErrors.invalidTransaction(
              'Options derived from operations do not match provided metadata options'
            )
          );
      }

      // Validate the sender address matches the public key.
      const derivedAddress = getAddressFromPublicKey(normalizedPublicKey, network);
      if (derivedAddress !== options.sender_address) {
        return reply
          .status(500)
          .send(
            MeshErrors.invalidPublicKey('The provided public key does not match the sender address')
          );
      }

      // Extract the fee amount from operations.
      const feeOperation = operations.find(op => op.type === 'fee');
      if (!feeOperation) {
        return reply.status(500).send(MeshErrors.invalidTransaction('Fee operation is required'));
      }
      if (feeOperation.account.address !== options.sender_address) {
        return reply
          .status(500)
          .send(
            MeshErrors.invalidTransaction('Fee payer address does not match the sender address')
          );
      }
      const feeAmount = BigNumber(feeOperation.amount.value);
      if (feeAmount.gte(0)) {
        return reply.status(500).send(MeshErrors.invalidTransaction('Fee amount must be negative'));
      }
      const fee = feeAmount.abs().toString();

      // Construct the unsigned transaction.
      let unsignedTx: StacksTransactionWire;
      switch (options.type) {
        case 'token_transfer': {
          unsignedTx = await makeUnsignedSTXTokenTransfer({
            recipient: options.recipient_address,
            amount: options.amount,
            fee,
            nonce: metadata.sender_account_info.nonce,
            publicKey: normalizedPublicKey,
            network,
            memo: options.memo,
          });
          break;
        }
        case 'contract_call': {
          const [contractAddress, contractName] = options.contract_identifier.split('.');
          unsignedTx = await makeUnsignedContractCall({
            contractAddress,
            contractName,
            functionName: options.function_name,
            functionArgs: options.args.map(arg => hexToCV(addHexPrefix(arg))),
            publicKey: normalizedPublicKey,
            fee,
            network,
          });
          break;
        }
        case 'contract_deploy': {
          unsignedTx = await makeUnsignedContractDeploy({
            contractName: options.contract_name,
            codeBody: options.source_code,
            clarityVersion: options.clarity_version,
            publicKey: normalizedPublicKey,
            fee,
            network,
          });
          break;
        }
      }

      // Compute and return the signing payload.
      const initialSighash = unsignedTx.signBegin();
      const sigHash = sigHashPreSign(
        initialSighash,
        AuthType.Standard,
        fee,
        metadata.sender_account_info.nonce
      );
      return reply.send({
        unsigned_transaction: addHexPrefix(serializeTransaction(unsignedTx)),
        payloads: [
          {
            account_identifier: { address: options.sender_address },
            address: options.sender_address,
            // Bare hex (no 0x): this is the signing payload, and Rosetta clients hex-decode it
            // directly (e.g. rosetta-sdk-go's encoding/hex rejects a 0x prefix). Matches the old
            // stacks-blockchain-api rosetta output. `unsigned_transaction`/`hash` stay 0x-prefixed.
            hex_bytes: removeHexPrefix(sigHash),
            signature_type: 'ecdsa_recovery',
          },
        ],
      });
    }
  );

  fastify.post(
    '/construction/combine',
    {
      schema: {
        body: ConstructionCombineRequestSchema,
        tags: ['Construction'],
        response: {
          200: ConstructionCombineResponseSchema,
          500: ErrorResponseSchema,
        },
      },
    },
    async (request, reply) => {
      const { unsigned_transaction, signatures } = request.body;

      const tx = deserializeTransaction(removeHexPrefix(unsigned_transaction));
      if (isSingleSig(tx.auth.spendingCondition)) {
        if (signatures.length !== 1) {
          return reply
            .status(500)
            .send(
              MeshErrors.invalidSignature(
                `Expected exactly 1 signature for single-sig transaction, got ${signatures.length}`
              )
            );
        }
        const sig = signatures[0];
        const signatureHex = removeHexPrefix(sig.hex_bytes);

        // Recoverable ECDSA signatures are 65 bytes (130 hex characters)
        if (signatureHex.length !== 130) {
          return reply
            .status(500)
            .send(
              MeshErrors.invalidSignature(
                `Invalid signature length: expected 130 hex characters (65 bytes), got ${signatureHex.length}`
              )
            );
        }
        // Rosetta signatures are [R|S|V] (recovery byte last); Stacks expects [V|R|S].
        tx.auth.spendingCondition.signature = createMessageSignature(
          reorderSignatureToVrs(signatureHex)
        );
      } else {
        // TODO: Multi-sig transaction combining
        return reply
          .status(500)
          .send(MeshErrors.notImplemented('Multi-sig transaction combining is not yet supported'));
      }

      return reply.send({
        signed_transaction: addHexPrefix(serializeTransaction(tx)),
      });
    }
  );

  fastify.post(
    '/construction/parse',
    {
      schema: {
        body: ConstructionParseRequestSchema,
        tags: ['Construction'],
        response: {
          200: ConstructionParseResponseSchema,
          500: ErrorResponseSchema,
        },
      },
    },
    async (request, reply) => {
      const { signed, transaction } = request.body;

      try {
        const txHex = removeHexPrefix(transaction);
        const decodedTx = codec.decodeTransaction(txHex);
        const senderAddress = decodedTx.auth.origin_condition.signer.address;
        const operations = await serializeDecodedTransactionOperations(decodedTx, config);

        const response: ConstructionParseResponse = {
          operations,
          ...(signed ? { account_identifier_signers: [{ address: senderAddress }] } : {}),
        };

        return reply.send(response);
      } catch (error) {
        const message = error instanceof Error ? error.message : 'Unknown error';
        return reply.status(500).send(MeshErrors.invalidTransaction(message));
      }
    }
  );

  fastify.post(
    '/construction/hash',
    {
      schema: {
        body: ConstructionHashRequestSchema,
        tags: ['Construction'],
        response: {
          200: TransactionIdentifierResponseSchema,
          500: ErrorResponseSchema,
        },
      },
    },
    async (request, reply) => {
      const { signed_transaction } = request.body;
      const tx = deserializeTransaction(removeHexPrefix(signed_transaction));
      return reply.send({
        transaction_identifier: {
          hash: addHexPrefix(tx.txid()),
        },
      });
    }
  );

  // Node-backed (broadcasts the transaction) — only served in online mode.
  if (config.mode === 'online') {
    const { rpcClient } = config;
    fastify.post(
      '/construction/submit',
      {
        schema: {
          body: ConstructionSubmitRequestSchema,
          tags: ['Construction'],
          response: {
            200: TransactionIdentifierResponseSchema,
            500: ErrorResponseSchema,
          },
        },
      },
      async (request, reply) => {
        const { signed_transaction } = request.body;
        try {
          const result = await rpcClient.request('POST', '/v2/transactions', {
            body: { tx: removeHexPrefix(signed_transaction) },
          });
          return reply.send({
            transaction_identifier: {
              hash: addHexPrefix(result),
            },
          });
        } catch (error) {
          const message = error instanceof Error ? error.message : 'Unknown error';
          return reply.status(500).send(MeshErrors.transactionBroadcastError(message));
        }
      }
    );
  }
};

function normalizeForUnorderedComparison(value: unknown): unknown {
  if (Array.isArray(value)) {
    return value
      .map(item => normalizeForUnorderedComparison(item))
      .sort((a, b) => JSON.stringify(a).localeCompare(JSON.stringify(b)));
  }
  if (value && typeof value === 'object') {
    const entries = Object.entries(value as Record<string, unknown>)
      .map(([key, entryValue]) => [key, normalizeForUnorderedComparison(entryValue)] as const)
      .filter(([, entryValue]) => entryValue !== undefined)
      .sort(([a], [b]) => a.localeCompare(b));
    return Object.fromEntries(entries);
  }
  return value;
}

function isUnorderedDeepStrictEqual(left: unknown, right: unknown): boolean {
  return isDeepStrictEqual(
    normalizeForUnorderedComparison(left),
    normalizeForUnorderedComparison(right)
  );
}

/**
 * Options with the memo removed, for the payloads options equality check. The memo is validated
 * and resolved separately because it may legitimately be present on only one side (e.g. supplied
 * via request metadata but not on the operations).
 */
function withoutMemo(options: ConstructionOptions): ConstructionOptions {
  if (options.type !== 'token_transfer' || options.memo === undefined) return options;
  const { memo: _memo, ...rest } = options;
  return rest;
}

function buildConstructionOptionsFromOperations(args: {
  operations: ConstructionOperation[];
  /** Request-level memo (legacy Rosetta style); operation-level memos take precedence. */
  memo?: string;
  maxFee?: string;
  suggestedFeeMultiplier?: number;
}): ConstructionOptions {
  const { operations, memo, maxFee, suggestedFeeMultiplier } = args;
  const maxFeeObject = maxFee ? { max_fee: maxFee } : undefined;
  const suggestedFeeMultiplierObject = suggestedFeeMultiplier
    ? { suggested_fee_multiplier: suggestedFeeMultiplier }
    : undefined;

  // A fee operation may be present (wallets often include the full operation set); it's ignored
  // here and the transfer options are derived from the remaining legs. `/construction/payloads`
  // still requires and reads the fee operation separately.
  const filteredOperations = operations.filter(op => op.type !== 'fee');

  // `contract_call` and `contract_deploy` are the only valid single operation types.
  // `token_transfer` requires two operations: one with a negative amount and one with a
  // positive amount. Other operation types are not allowed.
  switch (filteredOperations.length) {
    case 1: {
      const op = filteredOperations[0];
      switch (op.type) {
        case 'contract_call': {
          return {
            type: 'contract_call',
            sender_address: op.account.address,
            contract_identifier: op.metadata.contract_identifier,
            function_name: op.metadata.function_name,
            args: op.metadata.args,
            ...maxFeeObject,
            ...suggestedFeeMultiplierObject,
          };
        }
        case 'contract_deploy': {
          return {
            type: 'contract_deploy',
            sender_address: op.account.address,
            contract_name: op.metadata.contract_name,
            clarity_version: op.metadata.clarity_version,
            source_code: op.metadata.source_code,
            ...maxFeeObject,
            ...suggestedFeeMultiplierObject,
          };
        }
        default:
          throw new Error('Token transfers require two operations');
      }
    }
    case 2: {
      const [op1, op2] = filteredOperations;
      if (op1.type !== 'token_transfer' || op2.type !== 'token_transfer') {
        throw new Error('Token transfer operations cannot be combined with other operations');
      }
      let senderAddress: string | undefined;
      let recipientAddress: string | undefined;
      const value1 = BigNumber(op1.amount.value);
      const value2 = BigNumber(op2.amount.value);
      if (!value1.abs().eq(value2.abs())) {
        throw new Error('Token transfer operations require the same amount');
      }
      if (
        (value1.lt(0) && value2.lt(0)) ||
        (value1.gt(0) && value2.gt(0)) ||
        value1.eq(0) ||
        value2.eq(0)
      ) {
        throw new Error('Token transfers require one negative and one positive amount');
      }
      const memo1 = op1.metadata?.memo ?? undefined;
      const memo2 = op2.metadata?.memo ?? undefined;
      if (memo1 !== memo2) {
        throw new Error('Token transfer operations require the same memo');
      }
      if (value1.lt(0)) {
        senderAddress = op1.account.address;
        recipientAddress = op2.account.address;
      } else {
        senderAddress = op2.account.address;
        recipientAddress = op1.account.address;
      }
      return {
        type: 'token_transfer',
        sender_address: senderAddress,
        recipient_address: recipientAddress,
        amount: value1.abs().toString(),
        memo: memo1 ?? memo,
        ...maxFeeObject,
        ...suggestedFeeMultiplierObject,
      };
    }
    default:
      throw new Error('Invalid operation count: must be 1 or 2');
  }
}